# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Finanzas & Salud content site
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Article {
  id: number;
  slug: string;
  lang: string;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  publishedAt: string;
  imageUrl: string;
  tags: string[];
  featured: boolean;
}

export interface ArticleListResponse {
  articles: Article[];
  total: number;
}

export interface Category {
  id: number;
  slug: string;
  name: string;
  lang: string;
  articleCount: number;
}

export interface CategoryListResponse {
  categories: Category[];
}

export interface SiteStats {
  totalArticles: number;
  financeCount: number;
  healthCount: number;
  languages: string[];
}

export type NewsletterInputLang = typeof NewsletterInputLang[keyof typeof NewsletterInputLang];


export const NewsletterInputLang = {
  es: 'es',
  en: 'en',
  pt: 'pt',
} as const;

export interface NewsletterInput {
  email: string;
  lang?: NewsletterInputLang;
}

export interface NewsletterSubscription {
  id: number;
  email: string;
  /** @nullable */
  lang?: string | null;
  subscribedAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListArticlesParams = {
lang?: ListArticlesLang;
category?: ListArticlesCategory;
featured?: boolean;
limit?: number;
offset?: number;
};

export type ListArticlesLang = typeof ListArticlesLang[keyof typeof ListArticlesLang];


export const ListArticlesLang = {
  es: 'es',
  en: 'en',
  pt: 'pt',
} as const;

export type ListArticlesCategory = typeof ListArticlesCategory[keyof typeof ListArticlesCategory];


export const ListArticlesCategory = {
  finance: 'finance',
  health: 'health',
} as const;

export type GetArticleParams = {
lang?: GetArticleLang;
};

export type GetArticleLang = typeof GetArticleLang[keyof typeof GetArticleLang];


export const GetArticleLang = {
  es: 'es',
  en: 'en',
  pt: 'pt',
} as const;

export type ListCategoriesParams = {
lang?: ListCategoriesLang;
};

export type ListCategoriesLang = typeof ListCategoriesLang[keyof typeof ListCategoriesLang];


export const ListCategoriesLang = {
  es: 'es',
  en: 'en',
  pt: 'pt',
} as const;