# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Finanzas & Salud content site
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: articles
    description: Content articles
  - name: categories
    description: Article categories
  - name: newsletter
    description: Newsletter subscriptions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles
      description: Returns a paginated list of articles, optionally filtered by category, language, or featured status
      parameters:
        - name: lang
          in: query
          required: false
          schema:
            type: string
            enum: [es, en, pt]
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [finance, health]
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"

  /articles/{slug}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get a single article by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: lang
          in: query
          required: false
          schema:
            type: string
            enum: [es, en, pt]
      responses:
        "200":
          description: Article found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories with article counts
      parameters:
        - name: lang
          in: query
          required: false
          schema:
            type: string
            enum: [es, en, pt]
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListResponse"

  /stats:
    get:
      operationId: getSiteStats
      tags: [articles]
      summary: Get overall site stats
      responses:
        "200":
          description: Site statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteStats"

  /newsletter:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "201":
          description: Subscribed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscription"
        "409":
          description: Already subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Article:
      type: object
      required: [id, slug, lang, title, excerpt, content, category, publishedAt, imageUrl, tags, featured]
      properties:
        id:
          type: integer
        slug:
          type: string
        lang:
          type: string
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        publishedAt:
          type: string
        imageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean

    ArticleListResponse:
      type: object
      required: [articles, total]
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer

    Category:
      type: object
      required: [id, slug, name, lang, articleCount]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        lang:
          type: string
        articleCount:
          type: integer

    CategoryListResponse:
      type: object
      required: [categories]
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"

    SiteStats:
      type: object
      required: [totalArticles, financeCount, healthCount, languages]
      properties:
        totalArticles:
          type: integer
        financeCount:
          type: integer
        healthCount:
          type: integer
        languages:
          type: array
          items:
            type: string

    NewsletterInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        lang:
          type: string
          enum: [es, en, pt]

    NewsletterSubscription:
      type: object
      required: [id, email, subscribedAt]
      properties:
        id:
          type: integer
        email:
          type: string
        lang:
          type: ["string", "null"]
        subscribedAt:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string